% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vnorm}
\alias{vnorm}
\title{Run normalisation of reponse values}
\usage{
vnorm(
  ligandInfo,
  mean_GCaMP,
  pumps = NULL,
  exportXLSX = NULL,
  windowBefore = 5,
  windowAfter = 0,
  voltage = 5,
  matchByTime = FALSE
)
}
\value{
A dataframe
}
\description{
Run normalisation of reponse values as descibed in Jove et al 2020.
}
\author{
Veronica Jove
}
